Move the rpc handler code into a subpackage (climatevision.server)
