+++ src/climatevision/server/__init__.py
@@ -1,0 +1,4 @@
+# pyright: strict
+from .rpcs import GeneratorRpcs
+
+__all__ = ["GeneratorRpcs"]
